Progress on Falstad and toString methods
I worked more on the Falstad code and I also worked a bit on the
toString methods in some of the other files so that it's a bit easier to
debug what the data looks like. I'm having a bit of a hard time getting
the Falstad importer to connect the nodes currently.