Update readme's and gitignore
I cleared the core/*.dat files since those are generated from runs and
are not to be commited to master. They are outputs that are not
correctly put in the output folder, but because of testing, they may
still be important to be where they are. That should be fixed at a later
time.