Fix Vintage Story development env
I worked with @copygirl to resolve the issue - it appears that we do
need to in fact use ~/.xprofile (perhaps in addition to the ~/.bashrc?)
in order for VS Code to pick up on the environmental variables. I have
no idea how that would work on Windows, so they are on their own for the
time being.

I was able to test the game code, and all of our hooks in the API seem
to work, but my texture wasn't good (but whatever).