Fix launch file, add server launch, fix texture
I was working with Tyron from the Vintage Story devteam and for some
reason, VS Code does some weird things when you use the Integrated
Console. He fixed one of the bugs, but it still sends effectively
nonsense into the game console so we're going to use the Integrated
Terminal, which does not have this problem.

I added a new profile to use Mono to launch the server.

I fixed a texture path issue for our testing block.