Adds basic ore code to Minecraft
Added some basic copper ore to Minecraft. I was going to work on
ItemGroups to go with it but that turned into a dumpster fire quickly.

I also refactored the block loading in the main class to use a Kotlin
lambda function.

We have a list of other ores that we plan to add, but I do not have the
textures from my texture artist yet, those will be available in a few
days hopefully.

This also adds some code that does the actual population of the ores in
the world. I did not balance the copper at all, it was high enough to
make sure it worked.

When you break the ore block, it will drop a different item; this item
is not able to be placed again (because that doesn't make sense!).
I don't plan to allow Silk Touch to be able to pick up these blocks at
all, but I have not verified that.