Adds MatrixPrint and some tests for MNA
I added some new debug printing/formatting tools for dumping the
matricies from the circuit into the console in a readable format.

I also added a set of resistor tests which appears to show a bug with
the copying of data from the MNA unknowns matricies into the correct
nodes.