Changed capacitors to be based on charge
Falstad's implementation used the potential across a capacitor to determine what the equivalent current source should be. This works as long as no circuit topology changes swoop in and toss the voltage state stored in the nodes. When that occurs, it appears as if capacitors dump their charge into the void. By having charge be stored, this should not be able to happen.