Many changes with registration
I did a whole bunch of changes to the way that stuff is registered which
broke the mod in many different ways, the mod also does not build into a
distributable jar yet that can be run because of the darned Kotlin
intrinsics error (since the jar is not including the required Kotlin
libraries in it) probably due to the Gradle upgrade.

I also tried to reduce a bunch of the static enums and made new mappings
for that sort of thing to prevent us from writing the ore names in 10
places. It makes it a little easier, hopefully.