Added block and item registries. Set up infrastructure for cells (Nodes). Cells are created and managed by a cell tile entity, but they can also be created by the graph manager at startup (loading them from the disk). They exist in a CellGraph class, which holds the cells for the entire graph. To register a cell type, you need to implement the CellBase class and a Cell Provider, which requires has a factory method and a connection predicate, used to check if connections can be made for a certain direction. The cell then needs to be registered using the deferred register in the CellRegistry class. Then, the CellBlockBase class must be inherited from. Override the getCellProvider method to return the resource location from the registry. Also register it in the Block Registry class.
The tile entities will create and place the cell when they are placed. They will call the SetPlaced method on the cell class. The tile entities also store the graph ID to disk. When they get loaded (when the chunks get loaded) they will read the graph ID, get the graph, then get the cell instance using their position for lookup. They also call the methods to update the graph and connections for the cells, and the destroy method when they get broken by the player.